+++ src/component/Schema/Params/Element/assertVisible.js
@@ -1,6 +1,20 @@
 import { buildAssertionTpl } from "service/assert";
-import { AssertBoolean } from "../../Assert/AssertBoolean";
+import { AssertVisible } from "../../Assert/AssertVisible";
 import { renderTarget } from "service/utils";
+
+function visibleObjToString( obj ) {
+    if ( typeof obj.value !== "undefined" ) {
+      obj = {
+        isDisplayed: obj.value,
+        isVisible: obj.value,
+        isOpaque: obj.value,
+        isIntersecting: obj.value
+      };
+    }
+    return [ `${ obj.isDisplayed ? "" : "NOT " }displayed`, `${ obj.isVisible ? "" : "NOT " }visible`,
+    `${ obj.isOpaque ? "" : "NOT " } opaque`, `${ obj.isIntersecting ? "" : "NOT " }within the viewport` ]
+      .map( val => `\`${ val }\`` ).join( ", " );
+  }
 
 export const assertVisible = {
   template: ( command ) => buildAssertionTpl(
@@ -9,22 +23,20 @@ export const assertVisible = {
     `// Asserting that ${ command.target } element is visible`
   ),
 
-  toLabel: ({ assert }) => `(\`${ assert.value ? "visible" : "not visible" }\`)`,
+  toLabel: ({ assert }) => `(${ visibleObjToString( assert ) })`,
 
   toGherkin: ({ target, params, assert }) => `Assert that element \`${ target }\`
-    is \`${ assert.value ? "visible" : "not visible" }\` on the page`,
+    is ${ visibleObjToString( assert ) }`,
 
   commonly: "assert it is visible",
 
   assert: {
-    node: AssertBoolean,
+    node: AssertVisible,
     options: {
       textNode: "is visible"
     }
   },
-  description: `Asserts that the element is currently visible in the current viewport
-([intersecting](https://pptr.dev/#?product=Puppeteer&version=v1.15.0&show=api-elementhandleisintersectingviewport),
-styles \`display\` is not \`none\`, \`visible\` is not \`hidden\`, \`opacity\` is not \`0\`)`,
+  description: `Asserts that the element is currently visible`,
   params: [
   ],