+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - improvement: add page.assertConsoleMessage
 - improvement: add page.assertDialog
 - improvement: add page.closeDialog
+- improvement: extend target.assertVisible with options `displayed`, `visible`, `opaque`, `within the current viewport`
 - improvement: all assertions accept the opposite assumptions (equal/not equal, contain/ doesn't contain)
 - improvement: mixed type assertion simplified - one can assert only for string value. Exceptions are assertProperty/Attribute where additionally one can assert for boolean
 - improvement: Jest/Puppeteer project (export) has option to run in Firefox ([custom build](https://www.npmjs.com/package/puppeteer-firefox))